requirements.txt
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.3.22
psycopg2-binary==2.8.6
confi==0.0.4.1
PyYAML==5.3.1
covador==0.10.1

gunicorn==20.0.4

pytest==6.2.1
faker==5.0.2
